% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{calculate_hex_coords}
\alias{calculate_hex_coords}
\title{Hex Coordinates on NBA Basketball Court}
\usage{
calculate_hex_coords(shots, binwidths)
}
\value{
a \code{data.frame}
}
\description{
Get the coordinates of hexes to plot
in the basketball nba court
}
