% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{circle_points}
\alias{circle_points}
\title{Draw a circle}
\usage{
circle_points(center = c(0, 0), radius = 1, npoints = 360)
}
\value{
a \code{data.frame}
}
\description{
Function to draw a circle based on a center and radius
}
