% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dan_theme.R
\name{dan_theme}
\alias{dan_theme}
\title{dan_theme}
\usage{
dan_theme(ggp_object)
}
\arguments{
\item{ggp_object}{ggplot2 object}
}
\value{
ggplot2 custom theme
}
\description{
Minimal ggplot2 theme using the Poppins and Roboto Mono fonts
}
