% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{geom_nba_backboard}
\alias{geom_nba_backboard}
\title{NBA Basketball Backboard GGplot2 Geom Path}
\usage{
geom_nba_backboard(size_backboard = 1.3, ...)
}
\value{
a \code{list} of \code{ggproto} objects
}
\description{
Get the geom_paths to plot the basketball nba backboard
in a ggplot2 object
}
