% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{geom_nba_court}
\alias{geom_nba_court}
\title{NBA Basketball Court GGplot2 Geom Path}
\usage{
geom_nba_court(...)
}
\value{
a \code{ggproto} object
}
\description{
Get the geom_paths to plot the basketball nba court
in a ggplot2 object
}
