% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{nba_backboard_path}
\alias{nba_backboard_path}
\title{NBA Basketball Backboard Geom Path}
\usage{
nba_backboard_path()
}
\value{
a \code{ggproto} object
}
\description{
Get the geom_paths to plot the basketball nba backboard
in a ggplot2 object
}
