% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomodoro.R
\name{pomodoro}
\alias{pomodoro}
\title{Pomodoro timer}
\usage{
pomodoro(time_focus = 25, time_break = 5)
}
\arguments{
\item{time_focus}{focus time in minutes}

\item{time_break}{break time in minutes}
}
\description{
Pomodoro timer that runs as a background job
}
