% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{shots2hexbins}
\alias{shots2hexbins}
\title{Hexbins from Shots}
\usage{
shots2hexbins(
  shots,
  league_averages,
  binwidths = c(1.5, 1.5),
  min_radius_factor = 0.25,
  fg_diff_limits = c(-0.15, 0.15),
  fg_pct_limits = c(0.2, 0.7),
  pps_limits = c(0.5, 1.5)
)
}
\value{
a \code{data.frame}
}
\description{
Get the Hexbins stats from shots and league_averages
objects of a NBA shotchart API
}
