% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-utils.R
\name{theme_nba_court}
\alias{theme_nba_court}
\title{NBA Basketball Court Theme}
\usage{
theme_nba_court()
}
\value{
a \code{list} of \code{ggproto} objects
}
\description{
Style clean and minimal to plot the basketball nba court
}
